Support singular 'script' dirname convention
I can't find any documentation or any source that finds either of
'scripts' or 'script' to be more common in practice.

Absent any data, I lean only on anecdotal experience when I find
'script' to be more common in the wild than 'scripts'.

To further defend this, GitHub has [blogged](https://githubengineering.com/scripts-to-rule-them-all/)
their "Scripts to Rule Them All" pattern,
and for it they use 'script' singular.

This change adds 'script' (and 'script-win') as fallbacks if the plural
version does not exist.

This _should_ be a semver minor change.

Any users who may incidentally already have both scripts and script
directories, `scripts` will continue to be preferred (as long as it
exists). The singular 'script' is only used as a fallback when `scripts`
does _not_ exist.

The same is true for Windows users. `scripts-win` is preferred if both
`scripts-win` and `script-win` exist.

The _only_ users who may experience an issue are Windows users who do
not have a `scripts-win` directory but do have a `script-win` directory,
for some reason. In this case, what _would_ have fallen back to
`scripts` will now be preempted by `script-win`.