Tests that module script lookup falls back to script/
Refactors script-dirs tests to stub fs.existsSync instead of stubbing
process.cwd. (This removes the need to have fixtures of the various
script directory combinations.)

Additionally, this adds a test asserting that module script lookup
supports both scripts and script. (The latter being the default if
the former doesn't exist.)