Push the derivation of logLevel down a level
Originally, the log level derivation (respecting
verbose/silent/dryrun/quiet flags) happened at the CLI level. This was a
level too high because it meant that direct invocations of scripty would
not resolve the same log level (based on the other flags).

Now we do the log level resolution at the same point where all the
options are defaulted: within the optionify wrapper of scripty itself.

This way we still have access to distinguish the user-provided values vs
the defaults, and can apply the aliases and various supported flags.

Downside is that now the CLI needs to read in all the various flags to
pass them down, but that's an acceptable tradeoff.