Replace homegrown perms checking with fs.access
0.10 doesn't have fs.access and 0.10 was supported in the initial
scripty release.

We no longer support 0.10 so we can drop this homegrown module.

However, there is a quirk in that the old code specifically changed the
permission bits, but that didn't necessarily mean the file was actually
readable or executable.

For instance, a file with an +x but -r is not actually executable.
(Since the file needs read to actually execute.)
In addition, a non user-executable file with a group exec bit is not
actually executable by the node process.

Both of these scenarios are covered by fs.access.